kolejne procesy dodawanie auth i sprawdzanie isLoggIn
