+++ app/Opcje/utilities/helpers/onGoogleButtonPress.tsx
@@ -1,0 +1,22 @@
+import { GoogleAuthProvider, getAuth, signInWithCredential } from "@react-native-firebase/auth";
+import { GoogleSignin } from "@react-native-google-signin/google-signin";
+
+export async function onGoogleButtonPress() {
+  // Check if your device supports Google Play
+  await GoogleSignin.hasPlayServices({ showPlayServicesUpdateDialog: true });
+  // Get the users ID token
+  const signInResult = await GoogleSignin.signIn();
+
+  // Try the new style of google-sign in result, from v13+ of that module
+  const idToken = signInResult.data?.idToken;
+
+  if (!idToken) {
+    throw new Error("No ID token found");
+  }
+
+  // Create a Google credential with the token
+  const googleCredential = GoogleAuthProvider.credential(signInResult.data?.idToken);
+
+  // Sign-in the user with the credential
+  return signInWithCredential(getAuth(), googleCredential);
+}
